requirements.txt
pandas
deuces
openpyxl
psycopg2-binary
